Implement plan
This commit implements the planned changes.
