+++ src/components/layouts/MainLayout.tsx
@@ -1,0 +1,100 @@
+
+import { Button } from "@/components/ui/button";
+import { Link, Outlet, useNavigate } from "react-router-dom";
+import { Mail, Phone, MapPin, Github, Linkedin, Twitter } from "lucide-react";
+import { Input } from "@/components/ui/input";
+import { Textarea } from "@/components/ui/textarea";
+import { useToast } from "@/hooks/use-toast";
+
+const MainLayout = () => {
+  const navigate = useNavigate();
+  const { toast } = useToast();
+
+  const handleContactSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    toast({
+      title: "Message Sent",
+      description: "We'll get back to you as soon as possible.",
+    });
+  };
+
+  return (
+    <div className="min-h-screen flex flex-col bg-gradient-to-b from-gray-50 to-white">
+      {/* Navigation */}
+      <nav className="bg-white border-b border-gray-100">
+        <div className="container mx-auto px-4 py-6 flex justify-between items-center">
+          <Link to="/" className="text-2xl font-bold text-primary">
+            Open Office Survey
+          </Link>
+          <div className="hidden md:flex items-center space-x-8">
+            <Link to="/" className="text-gray-600 hover:text-primary">
+              Home
+            </Link>
+            <Link to="/features" className="text-gray-600 hover:text-primary">
+              Features
+            </Link>
+            <Button 
+              variant="outline" 
+              onClick={() => navigate('/login')}
+              className="font-medium"
+            >
+              Login
+            </Button>
+          </div>
+        </div>
+      </nav>
+
+      {/* Main Content */}
+      <main className="flex-1">
+        <Outlet />
+      </main>
+
+      {/* Footer with Contact Section */}
+      <footer className="bg-gray-900 text-white py-12">
+        <div className="container mx-auto px-4">
+          <div className="grid md:grid-cols-2 gap-12 max-w-4xl mx-auto mb-12">
+            <div className="space-y-6">
+              <h2 className="text-2xl font-bold mb-6">Get In Touch</h2>
+              <div className="flex items-center space-x-4">
+                <Mail className="h-6 w-6 text-primary" />
+                <span>contact@openofficesurvey.com</span>
+              </div>
+              <div className="flex items-center space-x-4">
+                <Phone className="h-6 w-6 text-primary" />
+                <span>+1 (555) 123-4567</span>
+              </div>
+              <div className="flex items-center space-x-4">
+                <MapPin className="h-6 w-6 text-primary" />
+                <span>123 Office Street, Business District</span>
+              </div>
+              <div className="flex space-x-4 pt-4">
+                <a href="#" className="text-gray-400 hover:text-primary">
+                  <Github className="h-6 w-6" />
+                </a>
+                <a href="#" className="text-gray-400 hover:text-primary">
+                  <Linkedin className="h-6 w-6" />
+                </a>
+                <a href="#" className="text-gray-400 hover:text-primary">
+                  <Twitter className="h-6 w-6" />
+                </a>
+              </div>
+            </div>
+            <form onSubmit={handleContactSubmit} className="space-y-4">
+              <Input placeholder="Your Name" className="bg-gray-800 border-gray-700" />
+              <Input type="email" placeholder="Your Email" className="bg-gray-800 border-gray-700" />
+              <Textarea placeholder="Your Message" className="h-32 bg-gray-800 border-gray-700" />
+              <Button type="submit" className="w-full">
+                Send Message
+              </Button>
+            </form>
+          </div>
+          <div className="border-t border-gray-800 pt-8 text-center text-gray-400">
+            <p>© {new Date().getFullYear()} Open Office Survey. All rights reserved.</p>
+          </div>
+        </div>
+      </footer>
+    </div>
+  );
+};
+
+export default MainLayout;
