+++ src/pages/Index.tsx
@@ -8,49 +8,13 @@ import {
   Shield, 
   Settings2, 
   Briefcase,
-  Mail,
-  Phone,
-  MapPin,
-  Github,
-  Linkedin,
-  Twitter
 } from "lucide-react";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
-import { useToast } from "@/hooks/use-toast";
 
 const Index = () => {
   const navigate = useNavigate();
-  const { toast } = useToast();
-
-  const handleContactSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    toast({
-      title: "Message Sent",
-      description: "We'll get back to you as soon as possible.",
-    });
-  };
 
   return (
-    <div className="min-h-screen bg-gradient-to-b from-gray-50 to-white">
-      {/* Navigation */}
-      <nav className="container mx-auto px-4 py-6 flex justify-between items-center">
-        <div className="text-2xl font-bold text-primary">Open Office Survey</div>
-        <div className="hidden md:flex items-center space-x-8">
-          <a href="#features" className="text-gray-600 hover:text-primary">Features</a>
-          <a href="/features" className="text-gray-600 hover:text-primary">Feature Details</a>
-          <a href="#how-it-works" className="text-gray-600 hover:text-primary">How It Works</a>
-          <a href="#contact" className="text-gray-600 hover:text-primary">Contact</a>
-          <Button 
-            variant="outline" 
-            onClick={() => navigate('/login')}
-            className="font-medium"
-          >
-            Login
-          </Button>
-        </div>
-      </nav>
-
+    <div>
       {/* Hero Section */}
       <div className="container mx-auto px-4 py-20">
         <div className="max-w-4xl mx-auto text-center">
@@ -158,90 +122,6 @@ const Index = () => {
           </div>
         </div>
       </section>
-
-      {/* Contact Section */}
-      <section id="contact" className="py-20 bg-white">
-        <div className="container mx-auto px-4">
-          <h2 className="text-3xl font-bold text-center mb-12">Get In Touch</h2>
-          <div className="grid md:grid-cols-2 gap-12 max-w-4xl mx-auto">
-            <div className="space-y-6">
-              <div className="flex items-center space-x-4">
-                <Mail className="h-6 w-6 text-primary" />
-                <span>contact@openofficesurvey.com</span>
-              </div>
-              <div className="flex items-center space-x-4">
-                <Phone className="h-6 w-6 text-primary" />
-                <span>+1 (555) 123-4567</span>
-              </div>
-              <div className="flex items-center space-x-4">
-                <MapPin className="h-6 w-6 text-primary" />
-                <span>123 Office Street, Business District</span>
-              </div>
-              <div className="flex space-x-4 pt-4">
-                <a href="#" className="text-gray-600 hover:text-primary">
-                  <Github className="h-6 w-6" />
-                </a>
-                <a href="#" className="text-gray-600 hover:text-primary">
-                  <Linkedin className="h-6 w-6" />
-                </a>
-                <a href="#" className="text-gray-600 hover:text-primary">
-                  <Twitter className="h-6 w-6" />
-                </a>
-              </div>
-            </div>
-            <form onSubmit={handleContactSubmit} className="space-y-4">
-              <Input placeholder="Your Name" />
-              <Input type="email" placeholder="Your Email" />
-              <Textarea placeholder="Your Message" className="h-32" />
-              <Button type="submit" className="w-full">
-                Send Message
-              </Button>
-            </form>
-          </div>
-        </div>
-      </section>
-
-      {/* Footer */}
-      <footer className="bg-gray-900 text-white py-12">
-        <div className="container mx-auto px-4">
-          <div className="grid md:grid-cols-4 gap-8 mb-8">
-            <div>
-              <h3 className="text-lg font-semibold mb-4">Open Office Survey</h3>
-              <p className="text-gray-400">Transform your workplace with meaningful feedback</p>
-            </div>
-            <div>
-              <h4 className="text-lg font-semibold mb-4">Features</h4>
-              <ul className="space-y-2 text-gray-400">
-                <li>Survey Management</li>
-                <li>Analytics Dashboard</li>
-                <li>Team Collaboration</li>
-                <li>Security</li>
-              </ul>
-            </div>
-            <div>
-              <h4 className="text-lg font-semibold mb-4">Resources</h4>
-              <ul className="space-y-2 text-gray-400">
-                <li>Documentation</li>
-                <li>API Reference</li>
-                <li>Privacy Policy</li>
-                <li>Terms of Service</li>
-              </ul>
-            </div>
-            <div>
-              <h4 className="text-lg font-semibold mb-4">Contact</h4>
-              <ul className="space-y-2 text-gray-400">
-                <li>Support</li>
-                <li>Sales</li>
-                <li>Partnerships</li>
-              </ul>
-            </div>
-          </div>
-          <div className="border-t border-gray-800 pt-8 mt-8 text-center text-gray-400">
-            <p>© {new Date().getFullYear()} Open Office Survey. All rights reserved.</p>
-            <p className="text-sm mt-2 opacity-50">Crafted with ♥ by THK</p>
-          </div>
-        </div>
-      </footer>
     </div>
   );
 };
